fix: chart function as a class
