+++ src/DatasetContent/types/index.ts
@@ -1,4 +1,3 @@
-import { type ChartData } from './Chart'
 import { type Stats } from './Stats'
 
-export type { ChartData, Stats }
+export type { Stats }
